% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PracticeUserInput.R
\docType{data}
\name{PracticeUserInput}
\alias{PracticeUserInput}
\title{A practice dataset that includes RNAseq data from a purified single cell type experiment}
\format{A data frame with 22462 rows and 18 variables:
\describe{
  \item{Mouse Gene Symbol}{The broadest category for the cell type where this gene was found to be specifically expressed}
  \item{A1}{The counts for the first sample of astrocytes}
  \item{A2}{The counts for the second sample of astrocytes}

}}
\source{
\url{https://sites.google.com/a/umich.edu/megan-hastings-hagenauer/}
}
\usage{
PracticeUserInput
}
\description{
This dataset contains RNAseq data from a purified single cell type experiment. Prior to usage, the data were log-transformed (base 2). Details can be found in: Zhang Y, Chen K, Sloan SA, Bennett ML, Scholze AR, O’Keeffe S, et al. An RNA-sequencing transcriptome and splicing database of glia, neurons, and vascular cells of the cerebral cortex. J Neurosci Off J Soc Neurosci. 2014 Sep 3;34(36):11929–47.
}
\keyword{datasets}

