% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sir_UnMixALot.R
\name{Sir_UnMixALot}
\alias{Sir_UnMixALot}
\title{Sir_UnMixALot}
\usage{
Sir_UnMixALot(userInput, dataColumns, geneColumn, species)
}
\arguments{
\item{userInput}{A dataframe containing the gene expression data for the samples (RNAseq or microarray data that has already been variance stabilized and received appropriate quality control to remove outlier samples and large-scale technical artifacts), including one column of gene symbols.  Defaults to Error.}

\item{dataColumns}{A vector containing the column numbers for the gene expression data in the dataframe. Defaults to Error.}

\item{geneColumn}{A single integer indicating the column number for the column that includes the gene symbols. Defaults to Error.}

\item{species}{A single character string indicating the species from which the data were derived, currently allows the values  "mouse", "Mouse", "human", or "Human". Defaults to "mouse".}
}
\value{
A list containing two data frames: PublicationSpecific_CellTypeIndex and AveragePrimary_CellTypeIndex. These data frames provide estimates for the relative balance of each cell type across samples. The first data frame provides estimates based on the cell type specific gene lists provided by particular publications ("cell type indices"), the second data frame averages each of the publication-specific cell type indices to create an average cell type index for each primary cell type.
}
\description{
This function uses Microarray or RNAseq data derived from heterogeneous brain cortical samples to estimate the relative balance of each cell type across samples. To do this, we use a database(CellTypeSpecificGenes_Master3) of genes that have been previously-indicated to have cell type specific expression in either the forebrain or cortex of humans or mice.
}
\examples{
Sir_UnMixALot(userInput=ZhangRNAseqData, dataColumns=c(2:18), geneColumn=1, species="mouse")
}
\keyword{RNAseq}
\keyword{brain}
\keyword{cell}
\keyword{cortex}
\keyword{deconvolution}
\keyword{matrix}
\keyword{microarray}
\keyword{type}

